# docs/api/swagger.json
{
   "swagger":"2.0",
   "info":{
      "version":"1.0.0",
      "title":"project1 API"
   },
   "basePath":"/v1/r",
   "host":"api.project1.io",
   "schemes":[
      "https"
   ],
   "consumes":[
      "application/json"
   ],
   "produces":[
      "application/json"
   ],
   "paths":{
      "/session/":{
         "post":{
            "parameters":[
               {
                  "name":"email",
                  "in":"formData",
                  "description":"email of the user to be login",
                  "required":true,
                  "schema":{
                     "properties":{
                        "email":{
                           "type":"string",
                           "format":"email"
                        },
                        "password":{
                           "type":"string"
                        }
                     }
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"the logined user",
                  "schema":{
                     "$ref":"User"
                  }
               },
               "default":{
                  "description":"error payload",
                  "schema":{
                     "$ref":"ErrorModel"
                  }
               }
            }
         },
         "get":{
            "responses":{
               "200":{
                  "description":"the current user",
                  "schema":{
                     "$ref":"User"
                  }
               },
               "default":{
                  "description":"error payload",
                  "schema":{
                     "$ref":"ErrorModel"
                  }
               }
            }
         },
         "delete":{
            "responses":{
               "200":{
                  "description":"logout successfully"
               },
               "default":{
                  "description":"error payload",
                  "schema":{
                     "$ref":"ErrorModel"
                  }
               }
            }
         }
      },
      "/user":{
         "post":{
            "responses":{
               "200":{
                  "description":"the created user",
                  "schema":{
                     "$ref":"User"
                  }
               },
               "default":{
                  "description":"error payload",
                  "schema":{
                     "$ref":"ErrorModel"
                  }
               }
            },
            "parameters":[
               {
                  "name":"cred",
                  "in":"body",
                  "description":"expected credential of new user",
                  "required":true,
                  "schema":{
                     "properties":{
                        "email":{
                           "type":"string",
                           "format":"email"
                        },
                        "password":{
                           "type":"string"
                        }
                     }
                  }
               }
            ],
            "operationId":"createUser",
            "tags":[
               "user"
            ]
         },
         "delete":{
            "parameters":[
               {
                  "name":"password",
                  "in":"query",
                  "description":"provide password to delete current user",
                  "required":true,
                  "schema":{
                     "properties":{
                        "password":{
                           "type":"string"
                        }
                     }
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"logout successfully"
               },
               "default":{
                  "description":"error payload",
                  "schema":{
                     "$ref":"ErrorModel"
                  }
               }
            }
         }
      }
   },
   "securityDefinitions":{
      "internalApiKey":{
         "type":"apiKey",
         "in":"header",
         "name":"api_key"
      }
   },
   "definitions":{
      "User":{
         "required":[
            "email"
         ],
         "properties":{
            "email":{
               "type":"string",
               "format":"email"
            },
            "id":{
               "type":"string",
               "format":"uuid"
            }
         }
      },
      "ErrorModel":{
         "required":[
            "message"
         ],
         "properties":{
            "message":{
               "type":"string"
            }
         }
      }
   }
}